# feat: Optimize iterableEquals function for better handling of Lists and other Iterables