## Status
READY

## Breaking Changes
NO

## Description
This PR optimizes the iterableEquals function to improve performance when comparing List types by using direct indexing. It also ensures efficient comparison for other Iterable types using iterators, while maintaining the function's compatibility with various iterable structures.


## Todos
- [x] Tests
- [x] Documentation
- [x] Examples

## Steps to Test or Reproduce
Steps to Test or Reproduce
The updated iterableEquals function improves performance by handling List types separately.

**Old Version:** Uses elementAt(i) for all Iterable types, which is efficient only for List. For other iterables like LinkedList or Queue, elementAt(i) can be inefficient (O(n)), potentially leading to O(n²) performance for large iterables.

**New Version:** Specifically optimizes List comparison by using direct indexing (O(1)), making it faster for lists. For other Iterable types, it uses iterators, ensuring linear time complexity (O(n)) for all cases. This avoids the performance bottleneck of elementAt(i).

In summary, the new version is more efficient, especially for List types, and ensures better performance for all Iterable structures by using the appropriate methods for each.


## Impact to Remaining Code Base
This PR will affect:

*