@felangel 2025-07-15
@Danilospano00 I just took a closer look and it looks like when you run in AOT, there is actually a regression. 

Before:
```
CollectionEquatable (static, large)
          total runs:     33 653
          total time:     2.0000  s
         average run:         59 μs
         runs/second:     16 949
               units:        100
        units/second:  1 694 915
       time per unit:     0.5900 μs
```

After:
```
CollectionEquatable (static, large)
          total runs:     20 008   
          total time:     2.0001  s
         average run:         99 μs
         runs/second:     10 101   
               units:        100   
        units/second:  1 010 101   
       time per unit:     0.9900 μs
```

`dart compile exe benchmarks/main.dart`